# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ticketing Gamification API
  version: 1.0.0
  description: REST API for ticketing + gamification. All protected routes under `/api/v1` require JWT (and `X-API-Key` if enabled).
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional. Required only when API key enforcement is enabled on the server.
  parameters:
    PaginationLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 200
      description: Max items to return (default 50, max 200)
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: invalid role
            details:
              description: Optional extra information (e.g., field errors)
              nullable: true
              oneOf:
                - type: object
                - type: array
                  items: {}
                - type: string
    DataResponse:
      type: object
      required: [data]
      properties:
        data: {}
    PagedResponse:
      allOf:
        - $ref: '#/components/schemas/DataResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
            limit:
              type: integer
    TicketStatus:
      type: string
      enum: [todo, in_progress, review, done]
    TicketPriority:
      type: string
      enum: [low, medium, high, urgent]
    TicketType:
      type: string
      enum: [feature, bug, chore]
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        username: { type: string }
        role: { type: string }
        avatarUrl: { type: string }
        badges:
          type: array
          items: { type: string }
        bio: { type: string }
...
    Ticket:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        type: { $ref: '#/components/schemas/TicketType' }
        assigneeId: { type: string, format: uuid, nullable: true }
        startDate: { type: string, format: date-time, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        history:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              text: { type: string }
              actorId: { type: string, format: uuid, nullable: true }
              timestamp: { type: string, format: date-time }
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        ticketId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        author: { type: string }
        text: { type: string }
        timestamp: { type: string, format: date-time }
...
paths:
...
  /api/v1/users:
    get:
      summary: List users
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
                      limit: { type: integer }
  /api/v1/users/me:
    get:
      summary: Current user profile
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    patch:
      summary: Update current user profile
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                bio: { type: string }
                avatarUrl: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /api/v1/users/{id}:
    get:
      summary: Get user
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    patch:
      summary: Update user role (admin only)
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
...
  /api/v1/tickets:
    get:
      summary: List tickets
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: projectId
          schema: { type: string, format: uuid }
        - in: query
          name: assigneeId
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/TicketStatus' }
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Ticket' }
                      limit: { type: integer }
    post:
      summary: Create ticket
      tags: [Tickets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, title, description, priority, type]
              properties:
                projectId: { type: string, format: uuid }
                title: { type: string }
                description: { type: string }
                priority: { $ref: '#/components/schemas/TicketPriority' }
                type: { $ref: '#/components/schemas/TicketType' }
                assigneeId: { type: string, format: uuid }
                startDate: { type: string, format: date-time }
                dueDate: { type: string, format: date }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ticket' }
  /api/v1/tickets/{id}:
    get:
      summary: Ticket detail
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ticket' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      summary: Delete ticket
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
  /api/v1/tickets/{id}/status:
    patch:
      summary: Update ticket status
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/TicketStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ticket' }
  /api/v1/tickets/{id}/comments:
    post:
      summary: Add ticket comment
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Comment' }
  /api/v1/comments/{commentId}:
    patch:
      summary: Update a comment (author only)
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Comment' }
    delete:
      summary: Delete a comment (author only)
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
...